Move FloatPoint3D
