+++ lib/FloatPoint3D.js
