+++ lib/PannerNode/ConeEffect.js
@@ -1,4 +1,4 @@
-var mathUtil = require('./mathUtil')
+var mathUtil = require('../mathUtil')
 
 class InvalidStateError extends Error {}
