+++ lib/PannerNode/PannerNode.js
@@ -1,7 +1,7 @@
 var AudioNode = require('../AudioNode')
   , AudioBuffer = require('../AudioBuffer')
   , BLOCK_SIZE = require('../constants').BLOCK_SIZE
-  , FloatPoint3D = require('./FloatPoint3D')
+  , FloatPoint3D = require('../FloatPoint3D')
   , DistanceEffect = require('./DistanceEffect')
   , ConeEffect = require('./ConeEffect')
   , PannerProvider = require('./PannerProvider')
